Hideous code, but works as a 0.0.0.0.0.1 version.
This commit removes once and for all the old V1 findNeighbour and all that. It now uses the git-like change accumulator with applyChanges and revertChanges. It also removes the example code from Renderer and starts adding implementation-relevant input forms and buttons.

Finally, this commit highlights the absolute necessity of an abstraction layer between nanogui/the renderer (will probably use RenderObjects or something like that) and the I/O operations that interface directly with both the Optimizer and the ThickSurface object.